Moving delegate classes into their individual places
