+++ include/godzilla/EssentialBC.h
@@ -5,8 +5,48 @@
 
 #include "godzilla/BoundaryCondition.h"
 #include "godzilla/Types.h"
+#include "godzilla/FunctionDelegate.h"
 
 namespace godzilla {
+
+namespace internal {
+
+struct EssentialBCFunctionMethodAbstract : public FunctionMethodAbstract {
+    virtual ErrorCode invoke_t(Int dim, Real time, const Real x[], Int nc, Scalar u[]) = 0;
+};
+
+template <typename T>
+struct EssentialBCFunctionMethod : public EssentialBCFunctionMethodAbstract {
+    EssentialBCFunctionMethod(T * instance,
+                              void (T::*method)(Real, const Real[], Scalar[]),
+                              void (T::*method_t)(Real, const Real[], Scalar[])) :
+        instance(instance),
+        method(method),
+        method_t(method_t)
+    {
+    }
+
+    ErrorCode
+    invoke(Int dim, Real time, const Real x[], Int nc, Scalar u[]) override
+    {
+        ((*this->instance).*method)(time, x, u);
+        return 0;
+    }
+
+    ErrorCode
+    invoke_t(Int dim, Real time, const Real x[], Int nc, Scalar u[]) override
+    {
+        ((*this->instance).*method_t)(time, x, u);
+        return 0;
+    }
+
+private:
+    T * instance;
+    void (T::*method)(Real, const Real[], Scalar[]);
+    void (T::*method_t)(Real, const Real[], Scalar[]);
+};
+
+} // namespace internal
 
 /// Essential boundary condition
 ///
